autogen-agent/requirements.txt
khaos-agent>=1.0.0
autogen-agentchat>=0.7.5
openai>=1.0
